requirements.txt
fastapi
uvicorn
PyMuPDF
requests
sentence-transformers
faiss-cpu
scikit-learn
google-generativeai